requirements-dev.txt
asgiref==3.4.1
bcrypt==3.2.0
boto3==1.17.108
botocore==1.20.108
cffi==1.14.5
cryptography==3.4.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.5
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-storages==1.11.1
djangorestframework==3.12.4
jmespath==0.10.0
paramiko==2.7.2
Pillow==8.3.1
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.6
